// src/app/models/end-of-day.ts
// Generated by https://quicktype.io

export interface EndOfDay {
  pagination: Pagination;
  data: Datum[];
}

export interface Datum {
  name?: string;
  market?: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
  adj_high: number;
  adj_low: number;
  adj_close: number;
  adj_open: number;
  adj_volume: number;
  symbol: string;
  exchange: string;
  date: string;
}

export interface Pagination {
  limit: number;
  offset: number;
  count: number;
  total: number;
}