// src/app/models/ticker.ts
// Generated by https://quicktype.io

export interface Ticker {
  name: string;
  symbol: string;
  has_intraday: boolean;
  has_eod: boolean;
  country: null;
  stock_exchange: StockExchange;
}

export interface StockExchange {
  name: string;
  acronym: string;
  mic: string;
  country: string;
  country_code: string;
  city: string;
  website: string;
}